refactor: move Workspace contents to a WorkspaceContent component
